+++ cmd/cortextool/pkg/client/alerts.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/pkg/errors"
 	log "github.com/sirupsen/logrus"
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
 const alertmanagerAPIPath = "/api/v1/alerts"