+++ cmd/cortextool/main.go
@@ -1,11 +1,13 @@
 package main
 
 import (
+	"fmt"
 	"os"
 
 	"gopkg.in/alecthomas/kingpin.v2"
 
 	"github.com/grafana/cortex-tools/pkg/commands"
+	"github.com/grafana/cortex-tools/pkg/version"
 )
 
 var (
@@ -18,7 +20,6 @@ var (
 )
 
 func main() {
-	kingpin.Version("0.1.3")
 	app := kingpin.New("cortextool", "A command-line tool to manage cortex.")
 	logConfig.Register(app)
 	alertCommand.Register(app)
@@ -26,6 +27,14 @@ func main() {
 	ruleCommand.Register(app)
 	pushGateway.Register(app)
 	loadgenCommand.Register(app)
+
+	app.Command("version", "Get the version of the cortextool CLI").Action(func(k *kingpin.ParseContext) error {
+		fmt.Print(version.Template)
+		version.CheckLatest()
+
+		return nil
+	})
+
 	kingpin.MustParse(app.Parse(os.Args[1:]))
 
 	pushGateway.Stop()