+++ cmd/cortextool/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## unreleased / master
 
+* [FEATURE] Added loki backend support for the rules commands, configurable with `--backend=loki` (defaults to cortex).
+
 ## v0.2.4
 
 * [BUGFIX] Fix alertmanager registration subcommand and path for the configuration API #72